requirements.txt
# Зависимости для PL/SQL Linter
Flask==2.3.2
PyYAML==6.0
python-dotenv==1.0.0